Agents/calendar_agent/requirements.txt
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-dateutil>=2.8.2
tzlocal>=4.3
requests>=2.0.0
